requirements.txt
# HFT FPGA Simulation - Python Dependencies

# Core
numpy>=1.24.0
pandas>=2.0.0

# Async networking (for live feed connections)
asyncio-websockets>=10.0
aiohttp>=3.8.0

# Order book data structures
sortedcontainers>=2.4.0

# FPGA simulation (optional)
myhdl>=0.11
cocotb>=1.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Profiling (for latency analysis)
py-spy>=0.3.0
memory-profiler>=0.60.0

# Optional: Live exchange connections
# ccxt>=4.0.0  # Crypto exchanges
# alpaca-trade-api>=3.0.0  # US stocks
